Handle hide/unhide sysfs
This is a mini-milestone - hiding the module
from /proc/modules is a must have, however we were
not hiding it from sysfs and that's a problem that
made the discovery of this lkm very trivial.

So we now have to handle kobjects, something
that, since this is not an actual driver, we should not
have to worry about, however sysfs is tightly related to kernel
objects, being the RAM mirror in the filesystem that serves
as communication between kernel and userland.

If we remove its sysfs entries it become impossible for
rmmod (libkmod) to unload the module as it will validate the entries
before unlinking the module from the kernel. So the solution
is to put back some sysfs attributes, only the essential ones,
for this module so we can safely hide/unhide/rmmod it, without
rebooting.